Created chunks.csv. More info in the description of the commit.
Created a .csv that contains information about the map.
For now it has only 1 "chunk", a group of 16 rows.
The 1st row has id of the chunk and the 4 neighbour chunks (-1 represents null).
The 15 following rows have the number of the "tile" that corresponds to the "block" (color) of the background.