Progress on gameState Refactoring (see description)
(Preliminary idea)
The Controller will have a GameState (Interface) that is implemented for example by menuGameState.
menuGameState has a set of 3 options each of which containing the text to display and the command itself for that option: new menuOption("Play", new playCommand())
and when that Play option is chosen the playCommand is executed.